backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
langgraph>=0.2.39,<0.4.0
httpx>=0.27.0,<0.29.0
pytest>=8.3.0,<9.0.0

